requirements.txt
# Core dependencies
rasterio>=1.3.0
gdal>=3.4.0
numpy>=1.21.0
asf-search>=6.0.0

# Web and data handling
requests>=2.28.0
urllib3>=1.26.0
beautifulsoup4>=4.11.0

# Configuration and serialization
pyyaml>=6.0
click>=8.0.0

# Date/time handling
python-dateutil>=2.8.0

# Scientific computing
scipy>=1.9.0
matplotlib>=3.5.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0